requirements.txt
Flask
requests
pyyaml
geoip2
